move KeepBodyAPIRoute
